Fix: user balance to float, add field status in bets

+++ betunfair/lib/betunfair/user.ex
@@ -3,7 +3,7 @@ defmodule Betunfair.User do
   import Ecto.Changeset
 
   schema "users" do
-    field :balance, :integer
+    field :balance, :float
     field :id_users, :string
     field :name, :string
